Converted UKS to dll Changed ModuleUKS from BEING the UKS to Referencing a UKS
