+++ BrainSimulator/Modules/ModuleUKSClauseDlg.xaml.cs
@@ -8,6 +8,7 @@ using System.Collections.Generic;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Media;
+using UKS;
 
 namespace BrainSimulator.Modules
 {
