+++ BrainSimulator/Modules/ModuleUKSQuery.cs
@@ -7,6 +7,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using UKS;
 
 namespace BrainSimulator.Modules
 {
