+++ BrainSimulator/Modules/ModuleUKSQueryDlg.xaml.cs
@@ -6,6 +6,7 @@
 
 using System.Collections.Generic;
 using System.Windows;
+using UKS;
 
 namespace BrainSimulator.Modules
 {
