+++ BrainSimulator/Modules/ModuleUKSStatementDlg.xaml.cs
@@ -10,6 +10,8 @@ using System.Linq;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Media;
+using UKS;
+
 
 namespace BrainSimulator.Modules
 {
