+++ BrainSimulator/Modules/Vision/ModuleVision.cs
@@ -12,6 +12,7 @@ using System.Windows.Media.Imaging;
 using System.Xml.Serialization;
 using System.Diagnostics;
 using static System.Math;
+using UKS;
 
 namespace BrainSimulator.Modules
 {
